+++ R/mod_caseBoxes.R
@@ -20,6 +20,12 @@ mod_caseBoxes_ui <- function(id) {
       column(3, shiny::uiOutput(ns("death"))),
       column(3, shiny::uiOutput(ns("recovered"))),
       column(3, shiny::uiOutput(ns("active")))
+    ),
+    fluidRow(
+      column(3, shiny::uiOutput(ns("new_confirmed"))),
+      column(3, shiny::uiOutput(ns("new_death"))),
+      column(3, shiny::uiOutput(ns("new_recovered"))),
+      column(3, shiny::uiOutput(ns("new_active")))
     )
   )
 }
@@ -37,9 +43,21 @@ mod_caseBoxes_server <- function(input, output, session, counts) {
              color = "white",
              background = case_colors[["confirmed"]])
   })
+  output$new_confirmed <- renderUI({
+    countBox("New",
+             counts()[["new_confirmed"]],
+             color = "white",
+             background = case_colors[["confirmed"]])
+  })
   output$death <- renderUI({
     countBox("Deaths",
              counts()[["deaths"]],
+             color = "white",
+             background = case_colors[["deaths"]])
+  })
+  output$new_death <- renderUI({
+    countBox("New",
+             counts()[["new_deaths"]],
              color = "white",
              background = case_colors[["deaths"]])
   })
@@ -49,9 +67,21 @@ mod_caseBoxes_server <- function(input, output, session, counts) {
              color = "white",
              background = case_colors[["recovered"]])
   })
+  output$new_recovered <- renderUI({
+    countBox("New",
+             counts()[["new_recovered"]],
+             color = "white",
+             background = case_colors[["recovered"]])
+  })
   output$active <- renderUI({
     countBox("Active",
              counts()[["active"]],
+             color = "white",
+             background = case_colors[["active"]])
+  })
+  output$new_active <- renderUI({
+    countBox("New",
+             counts()[["new_active"]],
              color = "white",
              background = case_colors[["active"]])
   })
@@ -61,7 +91,7 @@ countBox <- function(title, subtitle, color, background, width = "100%") {
   div(
     class = "count-box",
     shiny::h3(title),
-    shiny::p(subtitle),
+    shiny::p(formatC(subtitle, format = "f", big.mark = ",", digits  = 0)),
     style = sprintf(
       "color: %s; background-color: %s; width: %s;",
       color, background, width