+++ man-roxygen/ex-country-comparison.R
@@ -1,0 +1,26 @@
+if (interactive()) {
+  library(shiny)
+  long_title <- "Lorem ipsum dolor sit amet, consectetur adipisicing elit."
+  ui <- fluidPage(
+    mod_country_comparison_ui("country_comparison")
+  )
+  server <- function(input, output) {
+
+    orig_data <- reactive({
+      get_timeseries_full_data() %>%
+        get_timeseries_by_contagion_day_data()
+    })
+
+    orig_data_aggregate <- reactive({
+      orig_data_aggregate <- orig_data() %>%
+        aggregate_province_timeseries_data() %>%
+        add_growth_death_rate() %>%
+        arrange(Country.Region)
+      orig_data_aggregate
+    })
+
+    callModule(mod_country_comparison_server, "country_comparison", orig_data_aggregate = orig_data_aggregate)
+  }
+  runApp(shinyApp(ui = ui, server = server), launch.browser = TRUE)
+}
+