+++ NEWS.md
@@ -1,7 +1,16 @@
+### Covid19 1.1.6 (2020-04-06)
+
+- adding growth factor and lethality rate data to tables in country and country comparison tabs (#5)
+- adding growth factor and lethality rate plots to country comparison tab (#5)
+- adding comparison from 1000nd case to country comparison tab and country tab (#5)
+- fix NA in input data
+- add info to boxes #63
+- remove pop up with data source information (#63)
+
 ### Covid19 1.1.5 (2020-04-06)
 
 - fixed variables for map
--fixed albels for death rate plots
+- fixed albels for death rate plots
 
 ### Covid19 1.1.4 (2020-04-03)